Add common class and made new changes

+++ Controllers/ValuesController.cs
@@ -14,11 +14,15 @@ namespace PolicyDetails_WebApplication.Controllers
 {
     public class ValuesController : ApiController
     {
+        #region "Declare Common Parameters"
         private bool isValid = false;
         private string headerToken = string.Empty;
         private readonly string _strToken = "c06fc4189a5645e4a4fd480e8b1556e7";
         private readonly string _strAppName = "PolicyDetails";
         private readonly SunnyDBEntities _dbContext;
+        #endregion
+
+        #region "Constructor"
         public ValuesController() : base()
         {
             if (_dbContext == null)
@@ -26,7 +30,9 @@ namespace PolicyDetails_WebApplication.Controllers
                 _dbContext = new SunnyDBEntities();
             }
         }
+        #endregion
 
+        #region "Default methods"
         // GET api/values
         [System.Web.Http.Route("api/values/getValues")]
         public IEnumerable<string> Get()
@@ -54,7 +60,9 @@ namespace PolicyDetails_WebApplication.Controllers
         public void Delete(int id)
         {
         }
+        #endregion
 
+        #region "PolicyDetails Crud Opr"
         [HttpPost]
         [Route("savePolicyDetails")]
         public HttpResponseMessage savePolicyDetails([FromBody] PolicyData policyDetails)
@@ -63,19 +71,31 @@ namespace PolicyDetails_WebApplication.Controllers
             SuccessResponse sucessResponse = new SuccessResponse();
             try
             {
-                if (policyDetails == null)
+                headerToken = Request.Headers.GetValues(Enums.Enums.headerParams.headerToken.ToString()).First();
+                string APPName = Request.Headers.GetValues(Enums.Enums.headerParams.APPName.ToString()).First();
+                isValid = (headerToken == _strToken);
+                if (isValid && APPName == _strAppName)
                 {
-                    errorResponse.Status = "0";
-                    errorResponse.Message = "Request data not found";
-                    return Request.CreateResponse(HttpStatusCode.BadRequest, errorResponse);
+                    if (policyDetails == null)
+                    {
+                        errorResponse.Status = (int)Enums.Enums.statusCode.Error;
+                        errorResponse.Message = "Request data not found";
+                        return Request.CreateResponse(HttpStatusCode.BadRequest, errorResponse);
+                    }
+                    else
+                    {
+                        _dbContext.PolicyDatas.Add(policyDetails);
+                        _dbContext.SaveChanges();
+                        sucessResponse.Status = (int)Enums.Enums.statusCode.Success;
+                        sucessResponse.Message = "Policy data save successfuly";
+                        return Request.CreateResponse(HttpStatusCode.OK, sucessResponse);
+                    }
                 }
                 else
                 {
-                    _dbContext.PolicyDatas.Add(policyDetails);
-                    _dbContext.SaveChanges();
-                    sucessResponse.Status = "1";
-                    sucessResponse.Message = "Policy data save successfuly";
-                    return Request.CreateResponse(HttpStatusCode.OK, sucessResponse);
+                    errorResponse.Status = (int)Enums.Enums.statusCode.Error;
+                    errorResponse.Message = "Invalid API Key";
+                    return Request.CreateResponse(HttpStatusCode.NoContent, errorResponse);
                 }
             }
             catch (Exception ex)
@@ -92,30 +112,42 @@ namespace PolicyDetails_WebApplication.Controllers
             SuccessResponse sucessResponse = new SuccessResponse();
             try
             {
-                PolicyData _objPolicyData = _dbContext.PolicyDatas.FirstOrDefault(p => p.PolicyKey == paramPolicyData.PolicyKey);
-                if (_objPolicyData == null)
+                headerToken = Request.Headers.GetValues(Enums.Enums.headerParams.headerToken.ToString()).First();
+                string APPName = Request.Headers.GetValues(Enums.Enums.headerParams.APPName.ToString()).First();
+                isValid = (headerToken == _strToken);
+                if (isValid && APPName == _strAppName)
                 {
-                    errorResponse.Status = "0";
-                    errorResponse.Message = "Request data not found";
-                    return Request.CreateResponse(HttpStatusCode.BadRequest, errorResponse);
+                    PolicyData _objPolicyData = _dbContext.PolicyDatas.FirstOrDefault(p => p.PolicyKey == paramPolicyData.PolicyKey);
+                    if (_objPolicyData == null)
+                    {
+                        errorResponse.Status = (int)Enums.Enums.statusCode.Error;
+                        errorResponse.Message = "Request data not found";
+                        return Request.CreateResponse(HttpStatusCode.BadRequest, errorResponse);
+                    }
+                    else
+                    {
+                        _objPolicyData.ContractNumber = paramPolicyData.ContractNumber;
+                        _objPolicyData.CustomerCode = paramPolicyData.CustomerCode;
+                        _objPolicyData.RiskCommencementDate = Convert.ToDateTime(paramPolicyData.RiskCommencementDate);
+                        _objPolicyData.ProductName = paramPolicyData.ProductName;
+                        _objPolicyData.MaturityDate = paramPolicyData.MaturityDate;
+                        _objPolicyData.NextRenewalDue = Convert.ToDateTime(paramPolicyData.NextRenewalDue);
+                        _objPolicyData.SumAssuredAmount = Convert.ToDecimal(paramPolicyData.SumAssuredAmount);
+                        _objPolicyData.PremiumAmount = Convert.ToDecimal(paramPolicyData.PremiumAmount);
+                        _objPolicyData.ContractStatusCode = paramPolicyData.ContractStatusCode;
+                        _objPolicyData.PolicyStatus = paramPolicyData.PolicyStatus;
+                        _objPolicyData.ETLDate = Convert.ToDateTime(paramPolicyData.ETLDate);
+                        _dbContext.SaveChanges();
+                        sucessResponse.Status = (int)Enums.Enums.statusCode.Success;
+                        sucessResponse.Message = "Policy data save successfuly";
+                        return Request.CreateResponse(HttpStatusCode.OK, sucessResponse);
+                    }
                 }
                 else
                 {
-                    _objPolicyData.ContractNumber = paramPolicyData.ContractNumber;
-                    _objPolicyData.CustomerCode = paramPolicyData.CustomerCode;
-                    _objPolicyData.RiskCommencementDate = Convert.ToDateTime(paramPolicyData.RiskCommencementDate);
-                    _objPolicyData.ProductName = paramPolicyData.ProductName;
-                    _objPolicyData.MaturityDate = paramPolicyData.MaturityDate;
-                    _objPolicyData.NextRenewalDue = Convert.ToDateTime(paramPolicyData.NextRenewalDue);
-                    _objPolicyData.SumAssuredAmount = Convert.ToDecimal(paramPolicyData.SumAssuredAmount);
-                    _objPolicyData.PremiumAmount = Convert.ToDecimal(paramPolicyData.PremiumAmount);
-                    _objPolicyData.ContractStatusCode = paramPolicyData.ContractStatusCode;
-                    _objPolicyData.PolicyStatus = paramPolicyData.PolicyStatus;
-                    _objPolicyData.ETLDate = Convert.ToDateTime(paramPolicyData.ETLDate);
-                    _dbContext.SaveChanges();
-                    sucessResponse.Status = "1";
-                    sucessResponse.Message = "Policy data save successfuly";
-                    return Request.CreateResponse(HttpStatusCode.OK, sucessResponse);
+                    errorResponse.Status = (int)Enums.Enums.statusCode.Success;
+                    errorResponse.Message = "Invalid API Key";
+                    return Request.CreateResponse(HttpStatusCode.NoContent, errorResponse);
                 }
             }
             catch (Exception ex)
@@ -132,20 +164,32 @@ namespace PolicyDetails_WebApplication.Controllers
             SuccessResponse successResponse = new SuccessResponse();
             try
             {
-                PolicyData _objPolicyData = _dbContext.PolicyDatas.FirstOrDefault(p => p.PolicyKey == paramPolicy.PolicyKey);
-                if (_objPolicyData == null)
+                headerToken = Request.Headers.GetValues(Enums.Enums.headerParams.headerToken.ToString()).First();
+                string APPName = Request.Headers.GetValues(Enums.Enums.headerParams.APPName.ToString()).First();
+                isValid = (headerToken == _strToken);
+                if (isValid && APPName == _strAppName)
                 {
-                    errorResponse.Status = Enums.Enums.statusCode.Success.ToString();
-                    errorResponse.Message = "Policy details was not found";
-                    return Request.CreateResponse(HttpStatusCode.OK, errorResponse);
+                    PolicyData _objPolicyData = _dbContext.PolicyDatas.FirstOrDefault(p => p.PolicyKey == paramPolicy.PolicyKey);
+                    if (_objPolicyData == null)
+                    {
+                        errorResponse.Status = (int)Enums.Enums.statusCode.Success;
+                        errorResponse.Message = "Policy details was not found";
+                        return Request.CreateResponse(HttpStatusCode.OK, errorResponse);
+                    }
+                    else
+                    {
+                        _dbContext.PolicyDatas.Remove(_objPolicyData);
+                        _dbContext.SaveChanges();
+                        successResponse.Status = (int)Enums.Enums.statusCode.Success;
+                        successResponse.Message = "Record deleted successfuly";
+                        return Request.CreateResponse(HttpStatusCode.OK, successResponse);
+                    }
                 }
                 else
                 {
-                    _dbContext.PolicyDatas.Remove(_objPolicyData);
-                    _dbContext.SaveChanges();
-                    successResponse.Status = "1";
-                    successResponse.Message = "Record deleted successfuly";
-                    return Request.CreateResponse(HttpStatusCode.OK, successResponse);
+                    errorResponse.Status = (int)Enums.Enums.statusCode.Error;
+                    errorResponse.Message = "Invalid API Key";
+                    return Request.CreateResponse(HttpStatusCode.NoContent, errorResponse);
                 }
             }
             catch (Exception ex)
@@ -159,28 +203,30 @@ namespace PolicyDetails_WebApplication.Controllers
         public HttpResponseMessage Index(string Customer_Code)
         {
             ErrorResponse response = new ErrorResponse();
-            SuccessResponse successResponse = new SuccessResponse();
+            GetSuccessResponse successResponse = new GetSuccessResponse();
             try
             {
-                headerToken = Request.Headers.GetValues("headerToken").First();
-                string APPName = Request.Headers.GetValues("APPName").First();
+                DataTable _dtPolicyDetails = new DataTable();
+                headerToken = Request.Headers.GetValues(Enums.Enums.headerParams.headerToken.ToString()).First();
+                string APPName = Request.Headers.GetValues(Enums.Enums.headerParams.APPName.ToString()).First();
                 isValid = (headerToken == _strToken);
                 if (isValid && APPName == _strAppName)
                 {
-                    successResponse.PolicyDetails = this.getPolicyDataList(Customer_Code);
-                    successResponse.Status = "1";
+                    _dtPolicyDetails = Common.getPolicyDataList(Customer_Code);
+                    successResponse.PolicyDetails = Common.ConvertDataTable<PolicyData>(_dtPolicyDetails);
+                    successResponse.Status = (int)Enums.Enums.statusCode.Success;
                     if (successResponse.PolicyDetails != null && successResponse.PolicyDetails.Count > 0)
                         return Request.CreateResponse(HttpStatusCode.OK, successResponse);
                     else
                     {
-                        response.Status = "0";
+                        response.Status = (int)Enums.Enums.statusCode.Error;
                         response.Message = "Invalid customer code";
                         return Request.CreateResponse(HttpStatusCode.NoContent, response);
                     }
                 }
                 else
                 {
-                    response.Status = "0";
+                    response.Status = (int)Enums.Enums.statusCode.Error;
                     response.Message = "Invalid API Key";
                     return Request.CreateResponse(HttpStatusCode.NoContent, response);
                 }
@@ -190,53 +236,6 @@ namespace PolicyDetails_WebApplication.Controllers
                 return Request.CreateErrorResponse(HttpStatusCode.NoContent, ex); ;
             }
         }
-
-        public List<PolicyData_1> getPolicyDataList(string paramPolicyCode)
-        {
-            DataTable dt = new DataTable();
-            string cnnString = "Server= (local)\\SQLEXPRESS;DataBase=SunnyDB;trusted_connection=True;Encrypt=False;MultipleActiveResultSets=True;";
-
-            SqlConnection cnn = new SqlConnection(cnnString);
-            SqlCommand cmd = new SqlCommand();
-            cmd.Connection = cnn;
-            cmd.CommandType = System.Data.CommandType.StoredProcedure;
-            cmd.Parameters.Add(new SqlParameter("@CustoemerCode", paramPolicyCode));
-            cmd.CommandText = "SP_GetPolicyDetails";
-            cnn.Open();
-            SqlDataAdapter adapter = new SqlDataAdapter(cmd);
-            adapter.Fill(dt);
-            //object o = cmd.ExecuteScalar();
-            cnn.Close();
-            List<PolicyData_1> _objPolicyDataList = ConvertDataTable<PolicyData_1>(dt);
-
-            return _objPolicyDataList;
-        }
-        private static List<T> ConvertDataTable<T>(DataTable dt)
-        {
-            List<T> data = new List<T>();
-            foreach (DataRow row in dt.Rows)
-            {
-                T item = GetItem<T>(row);
-                data.Add(item);
-            }
-            return data;
-        }
-        private static T GetItem<T>(DataRow dr)
-        {
-            Type temp = typeof(T);
-            T obj = Activator.CreateInstance<T>();
-
-            foreach (DataColumn column in dr.Table.Columns)
-            {
-                foreach (PropertyInfo pro in temp.GetProperties())
-                {
-                    if (pro.Name == column.ColumnName)
-                        pro.SetValue(obj, dr[column.ColumnName], null);
-                    else
-                        continue;
-                }
-            }
-            return obj;
-        }
+        #endregion
     }
 }
